Update Taxon client
Added two additional functions to the API to support changes in https://github.com/kbase/taxon_api/tree/batch-data, these changes were tested locally.

Note that in the updated spec, the parameter object is named 'params'.  This will conflict with the functions internal usage of a 'params' variable.  In fact, using 'method' or 'options' or some other things might also cause things to break.  Not sure if we handled these cases before, but I know there are a lot of functions that use the name 'params'.